+++ src/app/handlers/command_handler.ts
@@ -13,15 +13,15 @@ import { joinedAtFixer } from '../feat-admin/joined_date_fixer/fix_joined_date.j
 import { shutdown } from '../feat-admin/shutdown/shutdown_process';
 import { uniqueChannelSettingsHandler } from '../feat-admin/unique_channel_settings/unique_channel_settings_hanlder.js';
 import { uniqueRoleSettingsHandler } from '../feat-admin/unique_role_settings/unique_role_settings_hanlder.js';
-import { anarchyRecruit } from '../feat-recruit/interactions/anarchy_recruit.js';
+import { anarchyRecruit } from '../feat-recruit/create_recruit/anarchy_recruit.js';
+import { eventRecruit } from '../feat-recruit/create_recruit/event_recruit.js';
+import { festRecruit } from '../feat-recruit/create_recruit/fest_recruit.js';
+import { regularRecruit } from '../feat-recruit/create_recruit/regular_recruit.js';
+import { salmonRecruit } from '../feat-recruit/create_recruit/salmon_recruit.js';
 import { buttonRecruit } from '../feat-recruit/interactions/commands/button_recruit.js';
 import { closeCommand } from '../feat-recruit/interactions/commands/close.js';
 import { otherGameRecruit } from '../feat-recruit/interactions/commands/other_game_recruit';
 import { privateRecruit } from '../feat-recruit/interactions/commands/private_recruit';
-import { eventRecruit } from '../feat-recruit/interactions/event_recruit.js';
-import { festRecruit } from '../feat-recruit/interactions/fest_recruit.js';
-import { regularRecruit } from '../feat-recruit/interactions/regular_recruit.js';
-import { salmonRecruit } from '../feat-recruit/interactions/salmon_recruit.js';
 import { handleIkabuExperience } from '../feat-utils/other/experience.js';
 import { handleFriendCode } from '../feat-utils/other/friendcode';
 import { handleHelp } from '../feat-utils/other/help.js';
